islands implementation cleanup and add to changelog
