+++ experiments/islands/pubspec.yaml
@@ -7,11 +7,11 @@ environment:
   sdk: '>=2.17.0 <3.0.0'
 
 dependencies:
-  jaspr: ^0.1.0
+  jaspr: ^0.2.0
 
 dev_dependencies:
   build_runner: ^2.1.4
   build_web_compilers: ^3.2.1
-  jaspr_builder: ^0.1.0
+  jaspr_builder: ^0.2.0
   lints: ^1.0.0
   webdev: ^2.7.10
