+++ packages/jaspr/CHANGELOG.md
@@ -3,6 +3,8 @@
 - Update to Dart 2.17
 - Added support for custom backend & server setup.
   Use the new `serveApp` method to wrap your app as a simple shelf handler
+- Added support for multiple apps on the client as well as island components.
+  With this jaspr supports the **islands architecture pattern**.
 
 # 0.1.5
