+++ packages/jaspr_builder/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+
+- Added builders for island components
+
 ## 0.1.0
 
 - Initial version.
